+++ webui/react/src/components/ActionSheet.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useRef } from 'react';
 import { CSSTransition } from 'react-transition-group';
 
 import css from './ActionSheet.module.scss';
-import Icon from './Icon';
+import Icon from '../tmp-shared/components/Icon/Icon';
 import Link, { Props as LinkProps } from './Link';
 
 export interface ActionItem extends LinkProps {
