+++ webui/react/src/components/Breadcrumb.stories.tsx
@@ -1,7 +1,7 @@
 import { Breadcrumb, Space } from 'antd';
 import React from 'react';
 
-import Icon from 'components/Icon';
+import Icon from 'tmp-shared/components/Icon/Icon';
 import Link from 'components/Link';
 import { paths } from 'routes/utils';
