+++ webui/react/src/components/DeterminedAuth.tsx
@@ -1,7 +1,7 @@
 import { Button, Form, Input } from 'antd';
 import React, { useCallback, useState } from 'react';
 
-import Icon from 'components/Icon';
+import Icon from 'tmp-shared/components/Icon/Icon';
 import Link from 'components/Link';
 import { StoreAction, useStoreDispatch } from 'contexts/Store';
 import { paths } from 'routes/utils';
