+++ webui/react/src/components/ResponsiveFilters.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import Dropdown, { Placement } from 'components/Dropdown';
 import useResize from 'hooks/useResize';
 
-import Icon from './Icon';
+import Icon from '../tmp-shared/components/Icon/Icon';
 import css from './ResponsiveFilters.module.scss';
 
 interface Props {
