+++ webui/react/src/components/TableFilterSearch.tsx
@@ -2,7 +2,7 @@ import { Button, Input } from 'antd';
 import { FilterDropdownProps } from 'antd/es/table/interface';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 
-import Icon from './Icon';
+import Icon from '../tmp-shared/components/Icon/Icon';
 import css from './TableFilterSearch.module.scss';
 
 interface Props extends FilterDropdownProps {
