+++ webui/react/src/pages/ModelVersionDetails/ModelVersionHeader.tsx
@@ -5,7 +5,7 @@ import React, { useCallback, useMemo, useState } from 'react';
 import Avatar from 'components/Avatar';
 import CopyButton from 'components/CopyButton';
 import DownloadModelModal from 'components/DownloadModelModal';
-import Icon from 'components/Icon';
+import Icon from 'tmp-shared/components/Icon/Icon';
 import InfoBox, { InfoRow } from 'components/InfoBox';
 import InlineEditor from 'components/InlineEditor';
 import Link from 'components/Link';
