+++ webui/react/src/pages/TrialDetails/Header/TrialHeaderLeft.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import Icon from 'components/Icon';
+import Icon from 'tmp-shared/components/Icon/Icon';
 import { paths } from 'routes/utils';
 import { getStateColorCssVar } from 'themes';
 import { ExperimentBase, TrialDetails } from 'types';
